requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
email-validator==2.2.0
python-multipart==0.0.9
jinja2==3.1.4
